Fix average displacement bug
Average displacements were filled with NaNs despite a sufficient number of high-confidence tracking points. It was caused by a mismatch between the frame on which users selected marker points and the frame used to group corner points. This commit ensures the code and the user are using the same frame.